refactor: layer aia chat componenet
